backend/requirements.txt
Django
djangorestframework
django-rest-auth
django-allauth
django-cors-headers
geopy
psycopg2
ipython